combine error handler hooks; deprecated.ts
